Remove jquery and bootstrap dependencies from tree.js
This is quite a big refactor, but helps make this much more standalone
(and the process of refactoring to use vanilla js access to a passed
document makes it much easier to unit test). I've tried to only use ES6
stuff (i.e. arrow functions) only where necessary, to keep the original
code more intact.

I've replaced the old bootstrap tooltip code with tippy.js, since we use
it on the entity graph already and it's a vanilla js replacement that's
almost identical.