Remove jQuery entirely and remove assets/javascripts
We've already removed it from all of our own JS, but this removes it
from rails (swapping for the newer rails_ujs which doesn't use it) and
swaps out the bootstrap JS for a vanilla-js version.

With this removed, we no longer need to use the asset pipeline for JS at
all, nor any of the gems which bundle JS libs, so I've removed the file
entirely.