Allow multiple relationship edges between nodes in graphs
When there are two different relationships, we need to show those
relationships separately, which our previous method of de-duping edges
by source/target id didn't allow for.

Instead, we dedupe edges by the Relationship object they represent and
just assign a random uuid for the id (because relationships don't have a
nice string id we could use instead).

This required a few changes to the specs for the entity graph because
we were using the ids as a nice shortcut to check which edges were
created, whereas now we need to check by relationship instead. To keep
consistency I also updated checks of nodes to use their entities.

I've also added some new specs to make the deduping feature of the
embedded Edge/Node classes clearer.

For OO-501