Fix bug with limiting controlled company links
We were showing the right links, but then counting the wrong number for
our 'more' count, because we weren't excluding ended relationships.

Note that we sort the displayed links by name, but we sort after we
limit it to 10 results from the DB. Elsewhere we've made the judgement
to load them all and then sort (e.g. in the paginated list on entity
pages) because it's critical to the function, so we're happy to take the
performance hit. (It's also mitigated there by only loading data for one
company). However, here it's a less critical part of the component
(these links are just supplementary info) and performance is more
important (we could be querying for 10 different entities in lists etc).