Store raw data as a String, not a Hash
Slovak data includes keys with . and potentially $ characters in, which
are illegal in MongoDB. I can't see or find any reliable way around this
which won't cause us problems later, or take us away from the primary
aim of the RawDataRecords, i.e. storing a faithful copy of the data we
saw.

This means we lose the ability to query the raw data, beyond regex
matching the text however. It's a hard trade off to make, but keeping
the data as faithful as possible wins out at the moment. It also means
we'll be able to be more faithful to data in non-json formats in the
future.