Convert DK import to use and save raw data
All in one commit for this source, as it's a bit of a pattern now. Note
I also changed the process to use Oj instead of the built-in JSON at the
same time, and I've added a series of new specs for the components of
the import that were untested before, e.g. the trigger.

As with other parts of this process, the import task is no longer needed
since all the metadata is in the db now.