Compress raw_data over 1MB in size
MongoDB allows 16MB which I thought was sufficiently massive that we'd
never need to worry about it as a limit. I was wrong, as there are
records in the DK data over 30MB in size. Luckily, they compress to more
like 4MB since JSON is so verbose.

Since we think these are the outliers, and I don't think much data will
ever be over 1MB, I've picked a much lower limit before using
compression