Show country flags on entity graph nodes
Cytoscape is not very flexible when it comes to rendering images, you can
only have a background image in a node. To get flags in we have to then
move the text label outside of the node. Once we've done this, we have a
new set of problems with the arrow endpoints not taking the label
padding and other spacing into account, but I've mitigated those as best
I can with a variety of spacing/padding/margin options.