Add data migration for EITI, UA and OO DataSources
We need all of these sources in the DB so that we can refer to them for
the Source metadata in the BODS export.

Note I've added a whole list of EITI sources. Whilst we collectively
refer to them as the EITI Pilot data, they are all individually linkable
and have unique document_ids, so I've had to make them separate. This
might highlight a flaw in our using the same model both for display and
metadata, but EITI data is small and a bit of an edge case anyway. It's
also highly likely that they came from different sources originally
(i.e. primary research done separately across different countries) and
so are best considered independent, even if we then combined and
standardised them.

For OO-242