Allow BodsMapper to output unknown people and unspecified ownerships
This puts together a lot of the previous code from this PR, allowing the
BodsMapper to output either an unknown personStatement or an unspecified
ownershipOrControlStatement as the data dictates.

Note that while we've done lots of work to make the relationships and
entities in these 'relationships' look like their real counterparts,
there are still some areas where they differ, like the lack of a
provenance. I've plugged this hole with a method to create provenance
from RawDataProvenances, which is a conversion we'll need in the future
anyway, but this is not complete.

Where I've significantly refactored things I've added a lot of specs for
the associated methods, but I've tried to limit it to only code I've
touched.

For OO-242