Allow BodsSerializer to output unknown people
Previously we skipped any relationship with an UnknownPersonsEntity or
a plain missing source/target entity. The latter are impossible (I
couldn't even figure out how to create them for a test without hitting
an issue in the code before the serializer) so I've removed the checks.
The former is now supported and so can be output as well.

For OO-242