Add system tests of BODS output
Using lib-cove-bods to validate the output. There's quite a bit of setup
needed to run the validator, hence why they can skipped if you don't set
the env var, but it seems to be the only way to reliably install it - I
had a simpler Docker-based version running, but it's not possible to use
that on CircleCI as our tests already run in a Docker container, so we
can't call out to run another one within them.

Note that I've replaced the existing request specs with these, they
cover the same tests but under the system test framework. I felt the
systems tests were a better location given the use of an external
validator (and the general end-to-endness of the specs). I've also
extended the tests to include a more extensive test of the actual output
as we don't really have anywhere else that covers it yet.

For OO-242