Make BODS statementIDs change when underlying data changes
Otherwise we're not outputting valid BODS, where a statement is meant to
be immutable, because we're changing the data without changing the ID.

Timestamps + IDs are used as a shortcut to figure out if an update has
happened, similar to how rails generates etags and cache keys.

I've followed the BODS guidance of adding a prefix, since our hashes are
particularly short and numeric, so more likely to collide.

Note that this is parallel to touching updated_at on Entities when
relationships change but does not require it. In fact, we use a separate
self_updated_at field on Entities which is not related to their
relationships.

For OO-554