Add a service object and worker for bulk BODS exports
I spent several weeks working (often in circles) on this in order to
achieve reasonable performance. There are three main issues:
- The sheer number of records we need to export (~20 million)
- The need to traverse relationships to output the right statements,
  resulting in a multiplication of database queries
- The need to map all of the loaded data to JSON, resulting in some
  expensive lookups for things like country names.

I tried both ways of iterating over the database (entity first and
relationship first) but settled on entities because of the need to output
companies with unknown owners. Outputting those is significantly simpler
if we start with every entity, because they naturally fall out.

I also considered multiple ways of storing the intermediate
data. Because it's an obvious case for parallelisation, we need a way to
recombine data at the end, and ensure the order. Again in no particular
order I tried:
- Using Redis to store finished statements and ids (won't scale without
  very expensive high-memory Redis instances)
- Pushing statements to S3 directly (too slow)
- Saving statements on disk and ids in Redis (means we can't run this
  directly on Heroku).

I went with the latter, which means that we have to run this job
somewhere with a local disk that persists for the length of the job
(as opposed to Heroku which resets every 24 hours). I intend to use EC2
for that purpose, which also allows us to fine-tune the capabilities to
get better performance in the future.

Finally, I spent a long time working on making it possible to
incrementally export data. This relies on accurate modification
timestamps, hence the previous commits to add those and ensure they
work. It also relies on some state being preserved between exports,
hence the new model and the statement-ids text file which we use to
bootstrap Redis during exports.

For OO-554