Paginate owned companies on entity page inside MongoDB
By dropping the requirement to sort by company name, and sorting solely
by relationship dates instead we can achieve a stable sort order and
hugely reduce the number or relationships and their associated companies
which have to be retrieved from the database.

We revisited the decision to sort by ended date during this, since the
combination of two dates (sorting in different directions) produced an
order that whilst superficially seems better, is actually much harder for
the user to reason about. Therefore we've gone with sorting by started
date only.

Note also that I haven't removed the use of RelationshipSorter
elsewhere. In particular it's use in ultimate_source_relationships and
other similar lists of people. There it's much more valuable to get
the original sort order, and since the numbers are much lower, it
doesn't (at the moment) have as big a performance impact.

For OO-594