Add a BODS export uploader service object
After trying various ways to upload data, this seems to be the simplest
because it relies on a local filesystem (removing lots of complexity
around copying/appending to files on S3) and simple commandline utilities
which stream and generally function a lot faster than implementing the
equivalent thing in Ruby.

The intention is that we'll run this on a dedicated EC2 box after the
export happens on the same box, meaning we'll have local files to play
with and a speedy connection to S3.

For OO-554