Eager load raw data provevances for relationships
This speeds up the BODS export where we need to use these provenances to
create source objects. Whilst it's not needed outside of BODS at the
moment, we will start using these records more often in the future, so
it doesn't hurt to have them ready then either.

For OO-554