Stop skipping existing child entities during PSC import
Now that we have incremental imports, this extra optimisation is less
useful and makes the code more complex than it needs to be. It also
results in us having a bigger stale data problem and makes that problem
harder to reason about. For those reasons, we're removing it.

For OO-567