Allow for empty people records as unknown owners from BODS
Unknown owners are modelled as a person with no name in BODS, which
breaks our assumptions about this only being modelled using
UnknownPersonsEntities from PSC statements. In addition, we'd allowed
for companies to have missing names, but not people, so we needed an
additional check and translation string for the latter case.