Add methods to query raw data provenance for entities
A variety of methods allowing us to build the raw data page of our
dreams. These all centre on having a specific entity and then being able
to query the imports, data sources and provenances that relate to them.

A lot of these are working around the issues we faced when having more
direct habtm relationships on things like data sources and imports (
those were unworkable due to the scale of the data). I've used some
memoisation where I know the values will be queried several times in the
same request, but I've made the decision not to memoise direct DB
queries because the database layer should already be doing this for us.

For OO-509