Remove Bucketeer and use our own S3 bucket
We get AWS credits, so there's no point paying for Bucketeer, and we
were only really using it to host our maintenance pages before this
export work started anyway.

Note I've replaced the generic app-wide access keys with ones specific
for the operation (which only give access to read/write the bucket). The
intention is that if we use S3 for anything else, we should generate
keys for that specific usage too.

For OO-554