Add a one-off migration to reimport ignored SK records
We didn't catch the fact when testing OO-480 that the changes to the
importer weren't sufficient to fix the issues in production with missing
data. Because we save all raw data, and only process new/changed records
in our imports, the existing records wouldn't be re-processed without
finding them and doing that explicitly.

This migration does that for SK data, finding the data which is not
attached to any record and re-importing it, so that it gets attached.

For OO-480