Download, extract and import UA data with one rake task
This combines the work we've done in
https://github.com/openownership/ua-edr-extractor and our existing
import into a single 'pipeline' command that can download the latest
data from the Ukraine government data portal, extract it and import the
resulting data.

Note that the new system tests require you to have installed
ua-edr-extractor locally, or they will be skipped.

For OO-615