Add libcovebods to requirements.txt
We only need it for testing, but it makes everything easier if all our
dependencies are in one place so I thought I'd tidy things up.

I also tidied up the way we check for both python dependencies in tests
since we're now clearer about where we can expect them to be.