Load companies from all jurisdictions in SK data, using a geocoder
This means that we can stop skipping foreign companies, but it also has
the knock-on effect that we should get more accurate data for all of the
companies in the data, because we're not mis-assigning any existing
foreign companies that happen to have a slovakian-looking postcode.

In testing this, we noticed in particular that there are over 1,000
Czech companies in the SK data which were wrongly assigned to SK.

Note that non-slovak companies get a registration number (Ico) on the
slovak system, but we shouldn't treat this as their customer number.

For OO-393