Change EntityResolver to do nothing when there's no jurisdiction_code
We never want to resolve with OC when we don't have a jurisdiction_code
because the results will be too broad to be useful. Existing importers
either always set a jurisdiction_code or have explicit checks to stop
themselves resolving when there isn't one. I'll remove those checks to
simplify them in future commits.

This came up in the light of adding foreign entities from SK data, where
we have to guess jurisdiction_codes from geocoding the address and so
might not find a jurisdiction_code at all. I didn't want to add another
check into that importer when we could make it irrelevant at source.

For OO-393