Merge duplicates found during SK import
We already do this for the UK import, and (I think) do it by default for
every import, but it's come to the fore for SK because we've started to
import foreign entities which overlap with existing data more. In
particular because we're re-assigning some existing entities to new
jurisdictions, which means we can match them to OC when we couldn't
before.

For OO-393