Output min/max for exact shares in BODS data
Because the standard says we should. I've also fleshed out the tests for
interest mapping more generally, because that was one of the areas where
we were lacking.

In addition, note the new error warning that our current hard-coded
exclusive_max/min settings are going to break when we get a BODS data
source that sets them. I realised a TODO wasn't really going to help us
remember this.

For OO-586