Record when relationships are indirect in DK data
This is useful information to show the user which we've been previously
ignoring. Note that we need to do slightly more work when displaying
inferred ultimate owner relationships, because we could have both a
single relationship which is actually indirect, or a known chain which
additionally includes some unknown indirect relationships.

For consistency purposes, I've also added the field to
Submissions::Relationship, but I've purposely not even set a default
because we're not making any assertions about whether the relationships
submitted are direct or indirect and we're not actively working on user
submission right now either so I don't have the bandwidth to think about
changing that.

For OO-159