Configure rails to use memcached cache_store
We didn't define any cache store in production at all, leaving Rails to
use the default file-based store, despite us paying for an running a
memcached server in production. In order to use Rack::Attack's
throttling we need a faster cache store than that, so it makes sense to
use memcached.

We already have a memcached setup in development too, so I've made sure
it gets started every time with docker (rather than requiring you to
start it manually) and configured the dev environment to use it if you
set the env vars. We already used an in-memory cache store in dev, but
this allows dev to better replicate prod which seems useful since we've
already done the hard work of making memcached available.