Rewrite memcached environment vars before rails env config
So that the generic ENV vars are available to use in config/environments.

Previously we only used memcached in the open corporates client, which
wouldn't run until the whole app was booted and the initializers had
run, but now we also want to use it for Rails' cache store and we
therefore need those env vars to exist much earlier in the app boot
process.