Add an exponential backoff to Rack::Attack throttling
We're still seeing a lot of traffic which isn't blocked and which is
overwhelming our database. With an exponential backoff (or an
approximation of it anyway) we're tracking good behaviour over a longer
period of time and punishing bad behaviour for longer too, which should
hopefully stop our abusive requestors for long enough.