Import BODS identifiers fully, instead of mapping to scheme => id
My initial BODS import tried to re-map BODS identifiers to a single key
value pair, more closely matching how we've created identifiers for other
data sources. This causes issues when we try to export those identiifiers
again, because we don't have enough information to recreate the
originals. Since being able to round-trip data is one of our key
measures for import/export quality, we need to preserve the original
structure. As we don't have any existing BODS data sources, this is easy
to change at this point, hence me rolling it into this ticket and not
providing any data migration, etc.

For OO-582