Export lots more identifiers in BODS data
After some discussion with the BODS team:
https://github.com/openownership/data-standard/issues/244

We decided that our previous conservative approach of only putting out
identifiers when we're sure they're officially verified could be improved.
The main reason for wanting to do this is to make our data export more
useful between versions, so that people can more accurately identify
companies/people. Secondarily, this should allow easier linking back to
the original data sources, and to OpenCorporates.

Now, we output the following:
- An OO register identifier for every entity
- An OpenCorporates identifier for every matched company
- The same official identifiers as before for companies declared in their
  home register. This has been slightly widened to include UK parent
  companies from the UK register.
- The same 'offical' beneficial owner identifiers for DK & SK as before,
  but with a warning that these are not actually using legit Org-Id
  scheme codes.
- All other source identifiers (self links, company numbers, names or
  combinations thereof) that the register stores.

For OO-582