Remove identifier caveats from /download, add OC matching caveats instead
Now we export lots of identifiers we can remove the general caveats and
offer more help on the historical data front. As suggested in the
related github ticket by Tim Davies, I've added some notes on the fact
that lots of fields actually come from OC, not the source.

For OO-582