Separate unknown entity names/explanations and remove tooltips
This is all very tangled, which is a strong sign that this part of the
code is a bit of a mess, but hopefully this big commit simplifies things
a little.

Essentially we want to display the reasons behind unknown entities being
'unknown' better than the current tooltips. People don't even know they
exist, and the info is very useful/interesting so should be more
prominent.

To do this I had to:
- Remove some things from the entity helpers which tried to gloss over
  the differences between Entities and UnknownPersonsEntities. There are
  less differences now, so less glossing needed, but I've also given
  some responsibilities back to individual views, rather than making
  tooltips magically appear in places.
- Add some checks for unknown entities to various 'title' partials, so
  that they could display the reasons for unknown-ness.
- Separate unknown reason codes from the textual description of those
  reasons at a model level. I removed some of my misguided copying of
  that text into unknown person entity names at the same time - that was
  done to make bods exports easier but was a poor data modelling choice.
- Tidy up how unknown entities got their names so that they were a)
  translatable and b) always came from the model itself, not other
  places that created them on the fly.

The way we create unknown entities on the fly is still a mess, and we'd
do well to replace it with a more BODS-like explicit record for unknown
owners, but that is not something for this ticket.

For OO-591