De-dupe statements during BODS export upload process
Despite our best efforts to de-dupe statement output, because we use
multiple independent workers and write their output to a list (not a
set), we can end up with duplicate statement ids in that list.

Therefore, when we combine the statements into a single file, we keep
another Redis set to de-dupe.

This is a bit of a kludge, it would be nicer if we could find an atomic
way to keep track of them and their order when we generate them, but I
couldn't figure one out through several hours of Redis docs reading so
I didn't want to waste any more time on it when this does solve the
issue.