Add a link to download entities as BODS into options bar
Whilst adding this, I also moved the graph page link into the options
bar and tidied up the code used for the icons therein, so that we use
font-awesome instead of custom SVGs. In order to be able to delete the
now redundant icons, I also had to tweak the 'report incorrect data'
button in the sidebar to use a FA icon as well.