New homepage
An updated homepage with some better examples of what the register does,
an explanation of what BO is and generally more of a sales pitch.

This broke lots of system tests, which were using a class wrapping the
search box to identify it. I've added a better id for it, so that we
aren't so reliant on the DOM structure in the future, and refactored the
code all the specs had into one shared helper.