Use local AWS credentials in BodsExportUploader
Rather than setting AWS credentials globally for the whole app to those
used by this service object, contain them within it. In a moment we're
going to add more credentials, so the global config will no longer make
sense.

This needed a few test tweaks to mock the appropriate extra objects
created and the http calls that makes in system tests.