Add back in cleaned up specs
Originally this repository contained some sample data, taken from the
various data sources and external systems we interact with, which we
used in various tests. When planning to open-source this repository we
received legal advice that we should remove this data.

To do so, we first invented data to replace it, using the formats and
structures of the existing sources, but generic values. You'll note that
all our tests use companies with names like "Example UK Company",
company numbers like 1234567 and dates which are all based around
starting points of 01/01/1950 (for people) or 01/01/2015 (for
companies).

We then rewrote the specs which used these files, or otherwise referenced
real people or companies, so that they didn't. This commit contains the
resulting tests and fixtures.