Paginate Elasticsearch response.records, not results
Paginating `response.results` breaks Kaminari 1.2+ because they check
for a `records` object on whatever you pass into `page_entries_info`.
Elasticsearch provides a stub for this method in a base class that
everything extends from and thus responds_to does indeed respond,
confusing Kaminari and breaking pagination.

We already load `response.records` when we display the results, so
there's nothing lost by giving it to Kaminari as well (at least as far
as I can see).

This is necessary to update kaminari to the latest version for an
(unrelated) security fix.