Output BODS identifiers for UK RLEs under a different scheme name
RLEs have two identifying pieces of information: a "self link" which
uniquely identifies them within the dataset and a "registration number"
which should be a globally unique company number (when used in
combination with the jurisdiction_code). We weren't outputting the
former when we had the latter for UK companies, but this didn't make
clear the difference in levels of verification/trust in the data.
Additionally, it resulted in lots of duplicate identifiers for some UK
companies, because despite them being different in our DB, the
identifiers got output as the same company number in BODS.

This adds a special case for these identifiers during output, resulting
in them generating two BODS identifiers for every one in our DB. One is
a "self link" identifier and is given the existing scheme name we've
already used for other self links (e.g. natural person owners), the
other is given a new scheme name to distinguish it from other company
numbers.

This means that we always get a self link identifier, *and* we get a
separate company number identifier showing the true provenance of that
company number.

In addition, the list of identifiers is now de-duped, because we can
still end up with multiple generated identifiers for a company that has
lots of merged records (e.g. where they're the RLE for lots of other
companies).