Document pausing of Ukraine EDR data imports
The data now contains a <beneficiaries> element where BO data should lie.
It's still not published as structured data, so we would need to adapt
our NLP-based ua-edr-extractor library to look in this field instead.
This will involve potentially significant development and particularly
testing which we're not well-placed to perform. This is before we get to
the issue of continuity between the data we already have (extracted from
a different data field) and whatever is in the new field.

For now, the best we can do is honestly document the situation for any
data users.