Force processing PSC data as UTF-8
I think previous versions of MongoDB/Mongoid were silently doing this
for us and then with our recent upgrade to 4.4 they stopped. This
resulted in encoding errors when we tried to bulk_write raw data.
Inspecting the lines we read from the zipped PSC data showed that some
are treated as UTF-8 whilst others are 'ASCII 8-BIT', i.e. binary data.
They're not, they're definitely always UTF-8, so we can force the
encoding.