Cache BODS JSON output for large entities
Some entities have very large amounts of BODS JSON, e.g. because they're
a person that owns a lot of companies. When this JSON is requested, it
causes a big spike in memory (because we have to load everything to
produce the JSON) and a large hit to the DB. Cumulatively, lots of
requests to these kinds of pages slows down our servers and causes
everyone's response times to increase.

We've solved this problem on HTML pages via pagination of specific lists
(e.g. merged people & owned companies), but we can't paginate BODS JSON
output in the same way because it's not coming out of a single database
query.

Instead, this is an attempt at improving the situation by caching the
expensive output. We have a memcached server going somewhat unused at
the moment, so we can use that to store the JSON which is causing us
problems.

This uses Rails built-in caching, which relies on the `#cache_key` method
on Entity including `updated_at`, and in turn on us ensuring `updated_at` is
touched when any related entity or relationship is updated. We did this
a while ago in order to generate BODS statement ids, so we can reuse it
here.

As a starting point, I've cached everything which has more than 20 BODS
statements, we can monitor cache usage and performance to see if this is
too small, but my suspicion is that there are quite few entities with
more than 20 statements, so we'll solve the problem once and for all
with this.