build: set RAILS_DEVELOPMENT_HOSTS
Without this, it's not possible to access the app within a VM without
tunnelling the connection or configuring NAT.

Also modify development environment to append. Later versions of Rails
read this env var automatically, so it may be that this is no longer
necessary once the app is upgraded (but the Docker bit still will be).