build: restore PATH to prioritise bin/
PATH was previously changed to give system paths a higher priority. This
was necessary because calling `yarn` would point to `bin/yarn`, which
didn't work when developing the build further up the chain. However,
everything is working with the current version of Yarn and Node.js, so
in fact the PATH can be restored to what I originally wrote.

This means that running the tests is as simple as

    docker compose exec web test

and entering a Rails console is as simple as

    docker compose exec web rails c

i.e. no need for `bin/test` or `bundle exec` or similar.