build: reduce number of volume mounts
Rails writes to log/ and tmp/ . Even though they are ignored in
.dockerignore , mounting the top-level repository would lead to the host
getting polluted by the container. Whilst this won't cause a problem in
many cases, it breaks the expectation that rebuilding a container will
result in a clean build. Not only that, it makes it harder to switch
branches productively, because it requires extra steps to do reliably.

There are a couple of main options, but both have disadvantages:

- mount additional volumes over the top of the volatile directories in
  the container (using named or unnamed volumes)
- mount only those specific paths from the host which frequently change
  during development (which leads to multiple volume mounts)

Here, the latter option is selected. Even if we wanted to select the
first option instead, that would be problematic, here: Node.js is
included and modules are installed by default in node_modules/ .
Mounting the top-level repo in the container as a single mount would
shadow the directory installed during the Yarn install phase in
Dockerfile , leading to it being missing.

Avoiding this would require either reworking things to go against the
usual Node.js expectation of dependencies being present in node_modules/
in the directory (which itself can become rather complex), or
potentially stacking mounts in some other way (which also gets rather
complex, and unless there is a guarantee of mount order: brittle).

To work around these considerations, and avoid additional unexpected
complexity, I've gone for simply choosing those parts that are usually
needed. However, here I've cleaned up the list and reduced it rather a
lot, as other files hardly ever change (or are perhaps no longer used).
Thus, if the rules for Rubocop need to be updated, for example, then
these won't auto-update within the container—but simply rebuilding the
container will update them. Hopefully, this is a good compromise between
complexity and functionality.