build: install Elasticsearch, Kibana
For local development and integration with other repos such as
register-transformer-psc , install a centralised Elasticsearch. It's
still possible to run multiple separate Elasticsearch services if
desired; however, it should not normally be needed. Connecting services
in this way allows for the datastores to be shared.

Install Elasticsearch 7.17.6 (current Elastic Cloud version).

Install Kibana 7.17.6 (matching Elasticsearch). This provides a UI for
exploring the Elasticsearch cluster and running queries (this is also as
providing by Elastic Cloud).

Connect web service to Elasticsearch, change example settings to expect
local cluster. This is likely more typical for development (and is far
safer). However, it's still fine to connect to remote Elasticsearch
clusters if desired.