build: optimise dev lib procedure to use env vars
When debugging multi-repo interaction, such as during investigation and
development of #163, it's necessary to override shared organisation
libraries and set up local Gems. Whilst the procedure I already put in
place for this works, it becomes somewhat inconvenient when switching
between multiple branches and versions of register-v2,
register-transformer-psc, and register-sources-bods, for example.

Optimise the dev lib procedure to use env vars:

- `DC_WEB_LIB_REGISTER_COMMON`
- `DC_WEB_LIB_REGISTER_SOURCES_BODS`
- `DC_WEB_LIB_REGISTER_SOURCES_OC`
- `DC_WEB_LIB_REGISTER_SOURCES_PSC`

When they are set, mount those libraries within the container.

On boot, detect that the libraries are mounted, and automatically
configure Bundler to mark them as local overrides.

Not only does this enable the paths of libraries on the host to be
customised, but also it allows services to be built with and without Gem
customisations simply by setting these env vars. Since they survive past
a service image rebuild, there are no additional manual steps needed.

This also decreases untracked changes pollution in the repo. Note that
there could still be some untracked changes in `Gemfile.lock`, however,
if pointing to a commit which the same as the locked one (which is
currently the case for register-v2, since the organisation libraries are
not up-to-date). Typically, these should not be checked in (unless doing
so as part of the development task).