[release-a7] build: rework Elasticsearch conn to use ELASTICSEARCH_URL
In order for non-root Elasticsearch users to be supported, don't
hard-code the `elastic` user within the connection config.

In addition to this, however, the config is simplified: instead of
supplying ELASTICSEARCH_PROTOCOL, ELASTICSEARCH_USER (which would be new),
ELASTICSEARCH_PASSWORD, ELASTICSEARCH_HOST, and ELASTICSEARCH_PORT, use
the ELASTICSEARCH_URL connection string format.

However, this environment variable is read by default anyway, meaning that
passing the environment variables when instantiating is not required. In
addition, the protocol defaults to HTTP, the user to `elastic`, the
password to empty, the host to `localhost`, and the port to `9200`,
meaning that all these defaults can be removed.

Also, since Elasticsearch `transport_options` are passed through to
Faraday, the default behaviour of SSL `verify=true` is used anyway for
HTTPS, meaning setting that shouldn't be necessary. Lastly, since
`log=false` is the default, that doesn't need to be specified, either.

This enables Elasticsearch configuration to be vastly simplified, as well
as additionally supporting non-root (i.e. non-`elastic`) users, whilst
defaulting to a config which Circle CI can use (i.e. `localhost:9200`).