[#187] add pages#download_latest
Register 1 offered a static link to the latest data file in AWS S3, as
well as links to individual files. Register 2 and friends doesn't write
this file into AWS S3. Instead, add a new route, which will redirect to
the latest AWS S3 file, offering the same behaviour as before, but at a
different URL.