[#188] uninstall various gems
These don't appear to be used. In the case of RSpec, there are no actual
Ruby tests anymore, seemingly since Register 2.

In doing this, I'm looking ahead to more regular upgrades, as well as
eventually a potentially upgrade to Rails 7. That's undecided, as yet,
but cleaning up will simplify the dependencies graph.

Some retained gems, particularly in groups, are unclear if they're
really needed, but I've compared them against a cleanly-generated
Rails 6.1.7.2 app, and taken that as guidance.

Note there are various gems that ideally perhaps might not be
dependencies; these are needed because of register-* internal libraries,
however, and the need to satisfy such within the project Gemfile for
internal gems published only via Git.