[#212] migrate content/ to config/data_sources.yml
In future, instead of updating Markdown in `content/`, update
`config/data_sources.yml`.

Note that the following datasources were not used, so the content wasn't
migrated:

- am-eiti-eregister
- am-multisector-register
- openownership-register

The previous content for these can be recovered from this commit if they
are needed in future, and added to `config/data_sources.yml`.