[#222] add alternate names to entities#show, twice
It's not clear to me whether only the Elasticsearch alternate names are
required, or whether the OpenCorporates API response should also be
surfaced. So, I've added both, in the different ways in which data from
each is currently represented (i.e. inconsistently). If necessary, one
or both of the alternate names blocks can easily be removed.

Note it wasn't actually mentioned in the ticket that this needed to be
done, but it seemed reasonable, given also what was done for adding
LEIs. Again, this can easily be reversed, if necessary.