[#215] refactor away PaginatedArray, using Register Common
Note that although most functionality between Register Common
`PaginatedArray` (previously in Sources BODS) and Register
`PaginatedArray` is the same, or with some extensions (in Register
Common), one important difference was found:

Register Common `offset_value` function subtracts 1 from the current
page, whereas Register does not. Looking back through the history of
each file, it isn't clear to me which is the correct behaviour. Also,
I'm not sure which entities would actually trigger this usage within raw
records, if any. So, I'm removing the duplicate code, knowing that a fix
may or may not turn out to be necessary for `offset_value`.